guess

+++ app/src/main/java/com/whzl/mengbi/model/entity/GameGuessBean.java
@@ -34,10 +34,10 @@ public class GameGuessBean {
         public String guessTheme;
         public String squareArgument;
         public String counterArgument;
-        public int squareArgumentFee;
-        public int counterArgumentFee;
-        public int squareOdds;
-        public int counterOdds;
+        public double squareArgumentFee;
+        public double counterArgumentFee;
+        public double squareOdds;
+        public double counterOdds;
         public String successArgument;
         public String status;
         public String closingTime;
